refactor : LottoSimulator 인터페이스 -> 클래스 변환
